+++ apps/frontend-control/postcss.config.js
@@ -1,5 +1,8 @@
 module.exports = {
   plugins: {
+    'postcss-import': {
+      addModulesDirectories: ['.', 'shared-components'],
+    },
     tailwindcss: {},
     autoprefixer: {},
     ...(process.env.NODE_ENV === 'production' ? { cssnano: {} } : {}),