+++ apps/frontend-manage/src/components/sessions/creation/MicroSessionWizard.tsx
@@ -9,7 +9,6 @@ import {
   FormikDateField,
   FormikSelectField,
   FormikTextField,
-  H3,
 } from '@uzh-bf/design-system'
 import dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
@@ -177,12 +176,34 @@ function MicroSessionWizard({
         }}
         onSubmit={onSubmit}
         isCompleted={isWizardCompleted}
+        editMode={!!initialValues}
         onRestartForm={() => {
           setIsWizardCompleted(false)
         }}
         onViewElement={() => {
           router.push(`/courses/${selectedCourseId}`)
         }}
+        workflowItems={[
+          {
+            title: 'Beschreibung',
+            tooltip:
+              'Geben Sie in diesem Schritt den Namen und die Beschreibung der Micro-Session ein.',
+          },
+          {
+            title: 'Einstellungen',
+            tooltip:
+              'Wählen Sie in diesem Schritt das Start- und Enddatum und nehmen Sie weitere Einstellungen vor.',
+            tooltipDisabled:
+              'Bitte überprüfen Sie zuerst Ihre Eingaben im vorherigen Schritt bevor Sie fortfahren.',
+          },
+          {
+            title: 'Fragen',
+            tooltip:
+              'Wählen Sie in diesem Schritt die Fragen für die Micro-Session aus.',
+            tooltipDisabled:
+              'Bitte überprüfen Sie zuerst Ihre Eingaben im vorherigen Schritt bevor Sie fortfahren.',
+          },
+        ]}
       >
         <StepOne validationSchema={stepOneValidationSchema} />
         <StepTwo validationSchema={stepTwoValidationSchema} courses={courses} />
@@ -215,30 +236,33 @@ interface StepProps {
 function StepOne(_: StepProps) {
   return (
     <>
-      <FormikTextField
-        required
-        autoComplete="off"
-        name="name"
-        label="Name"
-        tooltip="Der Name soll Ihnen ermöglichen, diese Micro-Session von anderen zu unterscheiden. Er wird den Teilnehmenden nicht angezeigt, verwenden Sie hierfür bitte den Anzeigenamen im nächsten Feld."
-        className={{ root: 'mb-1' }}
-        data-cy="insert-micro-session-name"
-      />
-      <FormikTextField
-        required
-        autoComplete="off"
-        name="displayName"
-        label="Anzeigename"
-        tooltip="Der Anzeigename wird den Teilnehmenden bei der Durchführung angezeigt."
-        className={{ root: 'mb-1' }}
-        data-cy="insert-micro-session-display-name"
-      />
+      <div className="flex flex-col w-full gap-4 md:flex-row">
+        <FormikTextField
+          required
+          autoComplete="off"
+          name="name"
+          label="Name"
+          tooltip="Der Name soll Ihnen ermöglichen, diese Micro-Session von anderen zu unterscheiden. Er wird den Teilnehmenden nicht angezeigt, verwenden Sie hierfür bitte den Anzeigenamen im nächsten Feld."
+          className={{ root: 'mb-1 w-full md:w-1/2' }}
+          data-cy="insert-micro-session-name"
+        />
+        <FormikTextField
+          required
+          autoComplete="off"
+          name="displayName"
+          label="Anzeigename"
+          tooltip="Der Anzeigename wird den Teilnehmenden bei der Durchführung angezeigt."
+          className={{ root: 'mb-1 w-full md:w-1/2' }}
+          data-cy="insert-micro-session-display-name"
+        />
+      </div>
 
       <EditorField
         label="Beschreibung"
         tooltip="Fügen Sie eine Beschreibung zu Ihrer Micro-Session hinzu, welche den Teilnehmern zu Beginn angezeigt wird."
         fieldName="description"
         data_cy="insert-micro-session-description"
+        showToolbarOnFocus={false}
       />
 
       <div className="w-full text-right">
@@ -255,7 +279,6 @@ function StepOne(_: StepProps) {
 function StepTwo(props: StepProps) {
   return (
     <div className="flex flex-col gap-2">
-      <H3 className={{ root: 'mb-0' }}>Einstellungen</H3>
       <div className="flex flex-row items-center gap-4 text-left">
         <FormikSelectField
           name="courseId"