+++ apps/frontend-manage/src/components/sessions/creation/SessionCreationBlock.tsx
@@ -9,9 +9,9 @@ import {
 } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { Ellipsis } from '@klicker-uzh/markdown'
-import { Button, Modal, NumberField, ThemeContext } from '@uzh-bf/design-system'
+import { Button, Modal, NumberField } from '@uzh-bf/design-system'
 import { move as RamdaMove } from 'ramda'
-import { useContext, useState } from 'react'
+import { useState } from 'react'
 import { useDrop } from 'react-dnd'
 import { twMerge } from 'tailwind-merge'
 
@@ -32,7 +32,6 @@ function SessionCreationBlock({
   move,
   replace,
 }: SessionCreationBlockProps): React.ReactElement {
-  const theme = useContext(ThemeContext)
   const [openSettings, setOpenSettings] = useState(false)
 
   const [{ isOver }, drop] = useDrop(
@@ -59,18 +58,17 @@ function SessionCreationBlock({
   )
 
   return (
-    <div
-      key={index}
-      className="flex flex-col p-2 border border-solid rounded w-52"
-    >
-      <div className="flex flex-row items-center justify-between">
-        <div className="font-bold" data-cy="block-container-header">
+    <div key={index} className="flex flex-col md:w-56">
+      <div className="flex flex-row items-center justify-between bg-slate-200 rounded py-1 px-2 text-slate-700">
+        <div className="" data-cy="block-container-header">
           Block {index + 1}
         </div>
-        <div className="flex flex-row gap-1 ml-2">
+        <div className="flex flex-row gap-1 text-xs">
           <Button
             basic
-            className={{ root: 'mx-1 disabled:hidden' }}
+            className={{
+              root: 'disabled:hidden hover:bg-primary-20 px-1',
+            }}
             disabled={numOfBlocks === 1}
             onClick={() => move(index, index !== 0 ? index - 1 : index)}
           >
@@ -80,7 +78,9 @@ function SessionCreationBlock({
           </Button>
           <Button
             basic
-            className={{ root: 'ml-1 mr-2 disabled:hidden' }}
+            className={{
+              root: 'disabled:hidden hover:bg-primary-20 px-1',
+            }}
             disabled={numOfBlocks === 1}
             onClick={() =>
               move(index, index !== numOfBlocks ? index + 1 : index)
@@ -90,32 +90,39 @@ function SessionCreationBlock({
               <FontAwesomeIcon icon={faArrowRight} />
             </Button.Icon>
           </Button>
+
           <Button
+            basic
+            onClick={() => setOpenSettings(true)}
+            className={{
+              root: 'px-1 sm:hover:text-primary hover:bg-primary-20',
+            }}
+          >
+            <Button.Icon>
+              <FontAwesomeIcon icon={faGears} />
+            </Button.Icon>
+          </Button>
+          <Button
+            basic
             onClick={() => remove(index)}
             className={{
-              root: 'w-6 flex justify-center text-white bg-red-500 rounded sm:hover:bg-red-600',
+              root: 'px-1  sm:hover:text-red-600',
             }}
             data={{ cy: 'delete-block' }}
           >
-            <FontAwesomeIcon icon={faTrash} />
-          </Button>
-          <Button
-            onClick={() => setOpenSettings(true)}
-            className={{
-              root: 'rounded w-6 flex justify-center',
-            }}
-          >
-            <FontAwesomeIcon icon={faGears} />
+            <Button.Icon>
+              <FontAwesomeIcon icon={faTrash} />
+            </Button.Icon>
           </Button>
         </div>
       </div>
-      <div className="flex flex-col flex-1 gap-1 my-2">
+      <div className="flex flex-col flex-1 my-2 md:overflow-y-auto">
         {block.titles.map((title, questionIdx) => (
           <div
             key={title}
-            className="flex flex-row text-xs border border-solid rounded bg-uzh-grey-20 border-uzh-grey-100"
+            className="flex flex-row items-center text-xs border-b border-solid border-slate-200 last:border-b-0 py-0.5"
           >
-            <div className="p-0.5 flex-1">
+            <div className="flex-1">
               <Ellipsis
                 // maxLines={2}
                 maxLength={40}
@@ -124,14 +131,11 @@ function SessionCreationBlock({
                 {title}
               </Ellipsis>
             </div>
-            <div className="h-full">
+            <div className="flex flex-row">
               <Button
                 basic
                 className={{
-                  root: twMerge(
-                    'flex flex-col justify-center h-1/2 px-2 disabled:hidden',
-                    theme.primaryBgHover
-                  ),
+                  root: 'flex flex-col justify-center disabled:hidden hover:bg-primary-20 px-1',
                 }}
                 disabled={block.questionIds.length === 1}
                 onClick={() => {
@@ -157,10 +161,7 @@ function SessionCreationBlock({
               <Button
                 basic
                 className={{
-                  root: twMerge(
-                    'flex flex-col justify-center h-1/2 px-2 disabled:hidden',
-                    theme.primaryBgHover
-                  ),
+                  root: 'flex flex-col justify-center disabled:hidden hover:bg-primary-20 px-1',
                 }}
                 disabled={block.questionIds.length === 1}
                 onClick={() => {
@@ -189,8 +190,11 @@ function SessionCreationBlock({
                 <FontAwesomeIcon icon={faArrowDown} />
               </Button>
             </div>
-            <div
-              className={`flex items-center px-2 text-white ${theme.primaryTextHover} bg-red-500 sm:hover:bg-red-600 rounded-r`}
+            <Button
+              basic
+              className={{
+                root: `px-1 hover:text-red-600`,
+              }}
               onClick={() => {
                 replace(index, {
                   ...block,
@@ -203,8 +207,10 @@ function SessionCreationBlock({
                 })
               }}
             >
-              <FontAwesomeIcon icon={faTrash} />
-            </div>
+              <Button.Icon>
+                <FontAwesomeIcon icon={faTrash} />
+              </Button.Icon>
+            </Button>
           </div>
         ))}
       </div>
@@ -212,7 +218,7 @@ function SessionCreationBlock({
         ref={drop}
         className={twMerge(
           'w-full text-center p-0.5 border border-solid rounded',
-          isOver && theme.primaryBg
+          isOver && 'bg-primary-20'
         )}
         data-cy="drop-questions-here"
       >