+++ apps/frontend-manage/src/pages/index.tsx
@@ -1,17 +1,23 @@
 import { useQuery } from '@apollo/client'
 import { GetUserQuestionsDocument } from '@klicker-uzh/graphql/dist/ops'
 import { useRouter } from 'next/router'
-import { useContext, useEffect, useMemo, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import useSortingAndFiltering from '../lib/hooks/useSortingAndFiltering'
 import { buildIndex, processItems } from '../lib/utils/filters'
 
 import TagList from '@components/questions/tags/TagList'
+import CreationButton from '@components/sessions/creation/CreationButton'
+import SessionCreation from '@components/sessions/creation/SessionCreation'
 import {
   faBoxArchive,
+  faChalkboardUser,
+  faGraduationCap,
   faMagnifyingGlass,
   faSort,
   faSortAsc,
   faSortDesc,
+  faUserGroup,
+  faUsersLine,
 } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
@@ -25,13 +31,14 @@ import { twMerge } from 'tailwind-merge'
 import Layout from '../components/Layout'
 import QuestionEditModal from '../components/questions/QuestionEditModal'
 import QuestionList from '../components/questions/QuestionList'
-import SessionCreation from '../components/sessions/creation/SessionCreation'
 
 function Index() {
   const router = useRouter()
-  const theme = useContext(ThemeContext)
 
   const [searchInput, setSearchInput] = useState('')
+  const [creationMode, setCreationMode] = useState<
+  undefined | 'liveSession' | 'microSession' | 'learningElement' | 'groupTask'
+  >(undefined)
   const [selectedQuestions, setSelectedQuestions] = useState<
     Record<number, boolean>
   >({})
@@ -61,7 +68,11 @@ function Index() {
   useEffect((): void => {
     router.prefetch('/sessions/running')
     router.prefetch('/sessions')
-  })
+
+    if (router.query.sessionId) {
+      setCreationMode(router.query.editMode as any)
+    }
+  }, [router])
 
   const index = useMemo(() => {
     if (dataQuestions?.userQuestions) {
@@ -140,30 +151,94 @@ function Index() {
       data={{ cy: 'homepage' }}
       className={{ children: 'pb-2' }}
     >
-      <div className="flex-none mb-4">
-        <SessionCreation
-          sessionId={router.query.sessionId as string}
-          editMode={router.query.editMode as string}
-        />
-      </div>
-
-      <div className="flex flex-row flex-1 gap-4 overflow-y-auto">
-        <div>
-          {dataQuestions && dataQuestions.userQuestions && (
-            <TagList
-              activeTags={filters.tags}
-              activeType={filters.type}
-              sampleSolution={filters.sampleSolution}
-              answerFeedbacks={filters.answerFeedbacks}
-              handleReset={handleReset}
-              handleTagClick={handleTagClick}
-              handleSampleSolutionClick={handleSampleSolutionClick}
-              handleAnswerFeedbacksClick={handleAnswerFeedbacksClick}
-              // handleToggleArchive={onToggleArchive}
-              // isArchiveActive={filters.archive}
-            />
-          )}
+      {typeof creationMode === 'undefined' && (
+        <div className="grid gap-1 mb-4 md:grid-cols-4 md:gap-2">
+          <CreationButton
+            icon={faUsersLine}
+            text="Live-Session erstellen"
+            onClick={() => {
+              setCreationMode('liveSession')
+            }}
+            data={{ cy: 'create-live-session' }}
+          />
+          <CreationButton
+            icon={faChalkboardUser}
+            text="Micro-Session erstellen"
+            onClick={() => {
+              setCreationMode('microSession')
+            }}
+            data={{ cy: 'create-micro-session' }}
+          />
+          <CreationButton
+            icon={faGraduationCap}
+            text="Lernelement erstellen"
+            onClick={() => {
+              setCreationMode('learningElement')
+            }}
+            data={{ cy: 'create-learning-element' }}
+          />
+          <CreationButton
+            icon={faUserGroup}
+            text="Gruppentask erstellen"
+            onClick={() => {
+              setCreationMode('groupTask')
+            }}
+            data={{ cy: 'create-group-task' }}
+            disabled
+          />
         </div>
+      )}
+      {creationMode && (
+        <div className="flex-none mb-4">
+          <SessionCreation
+            creationMode={creationMode}
+            closeWizard={() => {
+              router.push('/')
+              setCreationMode(() => undefined)
+            }}
+            sessionId={router.query.sessionId as string}
+            editMode={router.query.editMode as string}
+          />
+        </div>
+      )}
+
+      <div className="flex flex-col flex-1 gap-4 overflow-y-auto md:flex-row">
+        {dataQuestions && dataQuestions.userQuestions && (
+          <div>
+            <div className="hidden md:block">
+              <TagList
+                key={creationMode}
+                compact={!!creationMode}
+                activeTags={filters.tags}
+                activeType={filters.type}
+                sampleSolution={filters.sampleSolution}
+                answerFeedbacks={filters.answerFeedbacks}
+                handleReset={handleReset}
+                handleTagClick={handleTagClick}
+                handleSampleSolutionClick={handleSampleSolutionClick}
+                handleAnswerFeedbacksClick={handleAnswerFeedbacksClick}
+                // handleToggleArchive={onToggleArchive}
+                // isArchiveActive={filters.archive}
+              />
+            </div>
+            <div className="md:hidden">
+              <TagList
+                compact
+                key={creationMode}
+                activeTags={filters.tags}
+                activeType={filters.type}
+                sampleSolution={filters.sampleSolution}
+                answerFeedbacks={filters.answerFeedbacks}
+                handleReset={handleReset}
+                handleTagClick={handleTagClick}
+                handleSampleSolutionClick={handleSampleSolutionClick}
+                handleAnswerFeedbacksClick={handleAnswerFeedbacksClick}
+                // handleToggleArchive={onToggleArchive}
+                // isArchiveActive={filters.archive}
+              />
+            </div>
+          </div>
+        )}
         <div className="flex flex-col flex-1 w-full overflow-auto">
           {!dataQuestions || loadingQuestions ? (
             // TODO: replace by nice loader
@@ -232,10 +307,7 @@ function Index() {
                     setIsQuestionCreationModalOpen(!isQuestionCreationModalOpen)
                   }
                   className={{
-                    root: twMerge(
-                      'h-10 font-bold text-white',
-                      theme.primaryBgDark
-                    ),
+                    root: 'h-10 font-bold text-white bg-primary-80',
                   }}
                   data={{ cy: 'create-question' }}
                 >