+++ docs/src/components/common/TitleTextBlock.tsx
@@ -1,5 +1,5 @@
-import { ArrowRightIcon } from '@heroicons/react/solid'
-import React from 'react'
+import { faArrowRight } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { twMerge } from 'tailwind-merge'
 
 interface TitleTextBlockProps {
@@ -20,14 +20,18 @@ const TitleTextBlock = ({
     <div
       className={twMerge(
         className,
-        'relative p-4 md:p-8 border border-solid border-gray-200 rounded-lg shadow'
+        'relative rounded-lg border border-solid border-gray-200 p-4 shadow md:p-8'
       )}
     >
       <div className="mb-2 text-lg font-bold sm:text-xl">{title}</div>
       <div className="mb-10">{text}</div>
       {link && linkText && (
         <a href={link} className="absolute bottom-4">
-          <ArrowRightIcon className="h-5 align-text-bottom" /> {linkText}
+          <FontAwesomeIcon
+            icon={faArrowRight}
+            className="h-5 align-text-bottom mr-2"
+          />{' '}
+          {linkText}
         </a>
       )}
     </div>