+++ CHANGELOG.md
@@ -2,6 +2,49 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.4](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-beta.3...v3.0.0-beta.4) (2023-07-18)
+
+
+### Bug Fixes
+
+* **apps/frontend-*:** use relative paths to import shared components in globals.css ([f8e6d50](https://github.com/uzh-bf/klicker-uzh/commit/f8e6d5078a504f03418a2b65a571b33bbd43e698))
+
+## [3.0.0-beta.3](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-beta.2...v3.0.0-beta.3) (2023-07-18)
+
+## [3.0.0-beta.2](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-beta.1...v3.0.0-beta.2) (2023-07-18)
+
+
+### Other
+
+* **apps/frontend-manage:** improve styling of session components on session list with new collapsible ([#3667](https://github.com/uzh-bf/klicker-uzh/issues/3667)) ([3b8ff7d](https://github.com/uzh-bf/klicker-uzh/commit/3b8ff7d45359e8666ce7e010fc50ca2f528718f5))
+
+
+### Enhancements
+
+* **apps/frontend-manage:** introduce multistep wizard for session management, improve responsiveness of overall question pool layout ([#3765](https://github.com/uzh-bf/klicker-uzh/issues/3765)) ([cd19c7f](https://github.com/uzh-bf/klicker-uzh/commit/cd19c7f51749de64d15c643ec1b5a5a1d806ad89))
+* **docs:** upgrade docusaurus and packages, prepare versioning for v2/v3, update release announcement ([#3766](https://github.com/uzh-bf/klicker-uzh/issues/3766)) ([5f29b7f](https://github.com/uzh-bf/klicker-uzh/commit/5f29b7f50b3ebb2fa6aabf2e25316f9cfee40bb5))
+
+## [3.0.0-beta.1](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-beta.0...v3.0.0-beta.1) (2023-07-12)
+
+
+### Dependencies
+
+* add prettier-plugin-tailwindcss to pwa and control apps ([92b7ee1](https://github.com/uzh-bf/klicker-uzh/commit/92b7ee19a88d47937f6d7fb66f7070e58376cd07))
+
+## [3.0.0-beta.0](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.84...v3.0.0-beta.0) (2023-07-12)
+
+
+### Bug Fixes
+
+* **deploy:** add the cron token to the backend configmap ([69ce40a](https://github.com/uzh-bf/klicker-uzh/commit/69ce40ae4ab37c1983d94de50e8234331ec19afa))
+* **deploy:** ensure name of push cronjob is shorter than 52 char ([f70a0aa](https://github.com/uzh-bf/klicker-uzh/commit/f70a0aaab51060295d88897463c7f9abda4f24b7))
+* **deploy:** update supportEmail ([2f0db5d](https://github.com/uzh-bf/klicker-uzh/commit/2f0db5d3be2d2216f9e645083351c9d4542d092d))
+
+
+### Refactors
+
+* migrate all apps to new design-system theme approach ([#3757](https://github.com/uzh-bf/klicker-uzh/issues/3757)) ([3cbeb50](https://github.com/uzh-bf/klicker-uzh/commit/3cbeb50eae87f6d0f7a01615cb6bac8b21d19806))
+
 ## [3.0.0-alpha.84](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.83...v3.0.0-alpha.84) (2023-06-30)
 
 ## [3.0.0-alpha.83](https://github.com/uzh-bf/klicker-uzh/compare/v3.0.0-alpha.82...v3.0.0-alpha.83) (2023-06-30)